authentication_service/requirements.txt
asgiref==3.3.4
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
flake8==3.9.2
mccabe==0.6.1
psycopg2==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
python-dotenv==0.17.1
pytz==2021.1
sqlparse==0.4.1
